+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/home/giovanna/unb/E-a-Economia-WebApp/src/reportWebVitals.js":"1","/home/giovanna/unb/E-a-Economia-WebApp/src/index.js":"2","/home/giovanna/unb/E-a-Economia-WebApp/src/App.js":"3","/home/giovanna/unb/E-a-Economia-WebApp/src/routes.js":"4","/home/giovanna/unb/E-a-Economia-WebApp/src/services/auth.js":"5","/home/giovanna/unb/E-a-Economia-WebApp/src/containers/TelaLogin/index.js":"6"},{"size":362,"mtime":1606751595839,"results":"7","hashOfConfig":"8"},{"size":266,"mtime":1606760695473,"results":"9","hashOfConfig":"8"},{"size":109,"mtime":1606760664982,"results":"10","hashOfConfig":"8"},{"size":841,"mtime":1606759822381,"results":"11","hashOfConfig":"8"},{"size":332,"mtime":1606758981312,"results":"12","hashOfConfig":"8"},{"size":2809,"mtime":1606760532075,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1urk8mf",{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/home/giovanna/unb/E-a-Economia-WebApp/src/reportWebVitals.js",[],"/home/giovanna/unb/E-a-Economia-WebApp/src/index.js",[],"/home/giovanna/unb/E-a-Economia-WebApp/src/App.js",[],"/home/giovanna/unb/E-a-Economia-WebApp/src/routes.js",[],"/home/giovanna/unb/E-a-Economia-WebApp/src/services/auth.js",[],"/home/giovanna/unb/E-a-Economia-WebApp/src/containers/TelaLogin/index.js",["26"],{"ruleId":"27","severity":1,"message":"28","line":48,"column":11,"nodeType":"29","endLine":48,"endColumn":47},"jsx-a11y/alt-text","img elements must have an alt prop, either with meaningful text, or an empty string for decorative images.","JSXOpeningElement"]
