+++ src/App.js
@@ -1,0 +1,5 @@
+import React from "react";
+import Routes from "./routes";
+
+const App = () => <Routes />;
+export default App;
