+++ src/containers/TelaInicial/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function TelaInicial() {
+  return (
+    <div>TelaInicial</div>
+  );
+}
+
+export default TelaInicial;
