--- src/index.scss
@@ -1,16 +1,0 @@
-//material ui imports (font and icons)
-@import url('https://fonts.googleapis.com/css?family=Roboto:300,400,500,700&display=swap');
-@import url('https://fonts.googleapis.com/icon?family=Material+Icons');
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@100&display=swap');
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@500&display=swap');
-
-//default style to font
-* {
-  font-family: 'Poppins', sans-serif;
-}
-
-//default Material Ui button styles
-.MuiButtonBase-root {
-  width: 100% !important;
-  margin-top: 40px !important;
- }
