--- src/routes/HomeScreen/HomeScreen.js
@@ -1,10 +1,0 @@
-
-import React from 'react';
-
-function HomeScreen() {
-  return (
-    <div>HomeScreen</div>
-  );
-}
-
-export default HomeScreen;
